+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changelog
 1.7.0 (unreleased)
 ------------------
 
+- Add support for browser history navigation. Introduce ``ajax:path-target``,
+  ``ajax:path-action`` and ``ajax:path-event`` for this purpose.
+  [rnix, 2017-03-22]
+
 - ``bdajax.request`` supports ``method`` in options, defaults to ``GET``.
   [rnix, 2016-10-11]